Upgrade entity_id to UInt64
It seems paper uses varlong, especially when it gets late in the
server restart

This doesn't seem to have any negative downstream effects and fixes
the issues with civmc late in the day

+++ src/rosegold/packets/serverbound/entity_action.cr
@@ -14,7 +14,7 @@ class Rosegold::Serverbound::EntityAction < Rosegold::Serverbound::Packet
   end
 
   property \
-    entity_id : Int32,
+    entity_id : UInt64,
     action : Type,
     jump_boost : UInt8
